long int=>long long

+++ D7_RSA-keygen.cpp
@@ -2,45 +2,45 @@
 #include <stdio.h>
 #include <math.h>
 
-long int gcd(long int a, long int b);
-long int lcm(long int a, long int b);
-long int ExEuqlid(long int a, long int b);
+long long gcd(long long a, long long b);
+long long lcm(long long a, long long b);
+long long ExEuqlid(long long a, long long b);
 
 int main() {
 
     //異なる素数 p,q (p>q) //p*qが10桁程度までに設定
-    long int p=73; //ok
-    long int q=71;
+    long long p=73; //ok
+    long long q=71;
 
     if(p < q) { //to ensure p>q
-        long int tmp = p;
+        long long tmp = p;
         p = q;
         q = tmp;
     }
 
-    printf("P : %ld   Q : %ld\n",p,q);
+    printf("P : %lld   Q : %lld\n",p,q);
 
     //公開鍵 n
-    long int n = p * q;
+    long long n = p * q;
 
     //公開鍵 e
-    long int l = lcm(p-1, q-1); //最小公倍数 l
-    long int e = 2;
+    long long l = lcm(p-1, q-1); //最小公倍数 l
+    long long e = 2;
     while(gcd(l, e) != 1) {
         e++;
     }
 
     //秘密鍵 d
-    long int d = ExEuqlid(e, l);
+    long long d = ExEuqlid(e, l);
 
-    printf("N= %ld,  E= %ld,  D= %ld\n",n,e,d);
+    printf("N= %lld,  E= %lld,  D= %lld\n",n,e,d);
 
     return 0;
 }
 
-long int gcd(long int a, long int b) {
+long long gcd(long long a, long long b) {
 //最大公約数 Greatest Common Divisor
-     long int t;
+     long long t;
  
      while (b != 0) {
          t = a % b;  a = b;  b = t;
@@ -48,18 +48,18 @@ long int gcd(long int a, long int b) {
      return a;
 }
 
-long int lcm(long int a, long int b) {
+long long lcm(long long a, long long b) {
 //最小公倍数 Least Common Multiple
     return a * b / gcd(a, b);
 }
 
-long int ExEuqlid(long int a,long int b){
-    long int x, y, q;
-    long int x1 = 1, x2 = 0, x3;
-    long int y1 = 0, y2 = 1, y3;
-    long int result1 = b;
-    long int result2 = a;
-    long int result3;
+long long ExEuqlid(long long a,long long b){
+    long long x, y, q;
+    long long x1 = 1, x2 = 0, x3;
+    long long y1 = 0, y2 = 1, y3;
+    long long result1 = b;
+    long long result2 = a;
+    long long result3;
     while(1){
         q = result1 / result2;
         x3 = x1 - (q*x2);
